\name{survivalVignettes-package}
\alias{survivalVignettes-package}
\alias{survivalVignettes}
\docType{package}
\title{
\packageTitle{survivalVignettes}
}
\description{
\packageDescription{survivalVignettes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{survivalVignettes}
\packageIndices{survivalVignettes}
}
\author{
\packageAuthor{survivalVignettes}

Maintainer: \packageMaintainer{survivalVignettes}
}
\keyword{ package }
