\name{survivalVignettes-package}
\alias{survivalVignettes-package}
\alias{survivalVignettes}
\docType{package}
\title{
\packageTitle{survivalVignettes}
}
\description{
\packageDescription{survivalVignettes}
}

\author{
\packageAuthor{survivalVignettes}

Maintainer: \packageMaintainer{survivalVignettes}
}
\keyword{ package }
